refactor model to its own package
